add localstorage for endpoint
